{"types":"C","runs":34}

+++ Problem70.java
@@ -20,7 +20,7 @@ public class Problem70 extends GraphicsProgram implements ComponentListener {
 	private void drawColumns() {
 		for(int j = 0; j < N_Columns; j++){
 			double xCoord = (double) getWidth() / N_Columns;
-			GLine columnLine = new GLine(xCoord * j, 0, xCoord * j, getHeight());
+			GLine columnLine = new GLine((double) xCoord * j, 0, (double) xCoord * j, getHeight());
 			add(columnLine);
 		}
 	}
@@ -28,7 +28,7 @@ public class Problem70 extends GraphicsProgram implements ComponentListener {
 	private void drawRows() {
 		for(int i = 0; i < N_Rows; i++){
 			double yCoord = (double) getHeight() / N_Rows;
-			GLine rowLine = new GLine(0, yCoord * i, getWidth(), yCoord * i);
+			GLine rowLine = new GLine(0, (double) yCoord * i, (double) getWidth(), yCoord * i);
 			add(rowLine);
 		}
 	}
